refactor(tests): share the media tracker fixtures
`movie`, the Wire episode tree and the addon-registry wiring were each
built by hand in three modules. The ids had to agree across those files
for the episode-through-series assertions to mean anything, and that
kind of agreement stops holding quietly.

They move to `integration_test`, alongside `insert_test_source`.
